// apps/api/src/common/email/enums/email-category.enum.ts
export enum TransactionalEmailCategory {
  PASSWORD_RESET = 'PASSWORD_RESET',
  ACCOUNT_ONBOARDING = 'ACCOUNT_ONBOARDING',
  ORGANIZATION_APPROVAL = 'ORGANIZATION_APPROVAL',
  PROJECT_NOTIFICATION = 'PROJECT_NOTIFICATION',
  MEETING_NOTIFICATION = 'MEETING_NOTIFICATION',
  DELIVERABLE_NOTIFICATION = 'DELIVERABLE_NOTIFICATION',
  APPROVAL_DECISION = 'APPROVAL_DECISION',
  FINANCE_NOTIFICATION = 'FINANCE_NOTIFICATION',
  PURCHASE_NOTIFICATION = 'PURCHASE_NOTIFICATION',
  SALES_NOTIFICATION = 'SALES_NOTIFICATION',
  SYSTEM_SECURITY_ALERT = 'SYSTEM_SECURITY_ALERT',
}